requirements.txt
addict==2.4.0
basicsr==1.4.2
cElementTree==1.0.2-20050302
cog==0.9.24
coverage==7.6.1
elementtree==1.2.6-20050316
ffmpeg==1.4
gradio==4.44.1
gradio_client==1.3.0
huggingface_hub==0.24.6
ipdb==0.13.13
ipython==8.12.3
lxml==5.3.0
matplotlib==3.9.2
mock==5.1.0
networkx==3.3
onnxruntime==1.19.2
openai==1.51.0
opencv_python==4.10.0.84
pandas==2.2.3
Parser==0.1
Pillow==10.4.0
pip==24.2
protobuf==5.28.2
pycocotools==2.0.8
pydantic==2.9.2
Pygments==2.18.0
pytest==8.3.3
python-dotenv==1.0.1
pythran==0.16.1
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
runtests==0.0.28
scikit_learn==1.5.2
scipy==1.14.1
setuptools==72.1.0
Shapely==2.0.6
supervision==0.3.2
termcolor==2.4.0
timm==1.0.9
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transformers==4.44.2
yapf==0.40.2
